requirements.txt
bs4
requests
python-dotenv
smtplib